# Optimize findField operation heavily using a type map