--- a/datum_reader_test.go
+++ b/datum_reader_test.go
@@ -1,6 +1,7 @@
 package avro
 
 import (
+	"bytes"
 	"fmt"
 	"testing"
 )
@@ -265,4 +266,60 @@ func TestComplexOfComplexBinding(t *testing.T) {
 			}
 		}
 	}
+}
+
+func BenchmarkSpecificDatumReader_complex(b *testing.B) {
+	var dest complex
+	specificReaderBenchComplex(b, &dest)
+}
+
+func BenchmarkSpecificDatumReader_hugeval(b *testing.B) {
+	var dest struct {
+		complex
+		primitive
+		testRecord
+	}
+	specificReaderBenchComplex(b, &dest)
+}
+
+func specificReaderComplexVal() (Schema, []byte) {
+	schema, err := ParseSchemaFile("test/schemas/test_record.avsc")
+	if err != nil {
+		panic(err)
+	}
+	e := NewGenericEnum([]string{"A", "B", "C", "D"})
+	e.Set("A")
+	c := newComplex()
+	c.EnumField.Set("A")
+	c.FixedField = []byte{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15}
+	buf := testEncodeBytes(schema, c)
+	return schema, buf
+}
+
+func specificReaderBenchComplex(b *testing.B, dest interface{}) {
+	schema, buf := specificReaderComplexVal()
+	datumReader := NewSpecificDatumReader()
+	datumReader.SetSchema(schema)
+
+	b.ResetTimer()
+
+	for i := 0; i < b.N; i++ {
+		dec := NewBinaryDecoder(buf)
+		err := datumReader.Read(dest, dec)
+		if err != nil {
+			b.Fatal(err)
+		}
+	}
+}
+
+func testEncodeBytes(schema Schema, rec interface{}) []byte {
+	var buf bytes.Buffer
+	w := NewSpecificDatumWriter()
+	w.SetSchema(schema)
+	encoder := NewBinaryEncoder(&buf)
+	err := w.Write(rec, encoder)
+	if err != nil {
+		panic(err)
+	}
+	return buf.Bytes()
 }