--- a/datum_utils.go
+++ b/datum_utils.go
@@ -21,35 +21,50 @@ import (
 	"strings"
 )
 
-func fieldByTag(where reflect.Value, name string) reflect.Value {
-	elemType := where.Type()
-
-	for i := 0; i < where.NumField(); i++ {
-		field := where.Field(i)
-		if elemType.Field(i).Tag.Get("avro") == name {
-			return field
-		}
-	}
-
-	return reflect.Value{}
-}
-
 func findField(where reflect.Value, name string) (reflect.Value, error) {
 	if where.Kind() == reflect.Ptr {
 		where = where.Elem()
 	}
+	t := where.Type()
+	rm := reflectMap[t]
+	if rm == nil {
+		rm = reflectBuildRi(t)
+	}
+	if rf, ok := rm.names[name]; ok {
+		return where.FieldByIndex(rf), nil
+	}
+	return reflect.Value{}, fmt.Errorf("Field %s does not exist in %s", name, t.Name())
+}
 
-	field := fieldByTag(where, name)
-	if !field.IsValid() {
-		field = where.FieldByName(strings.ToUpper(name[0:1]) + name[1:])
-		if !field.IsValid() {
-			field = where.FieldByName(name)
+func reflectBuildRi(t reflect.Type) *reflectInfo {
+	rm := &reflectInfo{
+		names: make(map[string][]int),
+	}
+	for i := 0; i < t.NumField(); i++ {
+		f := t.Field(i)
+		if strings.ToLower(f.Name[:1]) != f.Name[:1] {
+			tag := f.Tag.Get("avro")
+			if tag != "" {
+				rm.names[tag] = f.Index
+			} else {
+				rm.names[f.Name] = f.Index
+				rm.names[strings.ToLower(f.Name[:1])+f.Name[1:]] = f.Index
+			}
 		}
 	}
 
-	if !field.IsValid() {
-		return reflect.Value{}, fmt.Errorf("Field %s does not exist", name)
+	// copy the map instead of dealing with locking
+	m := make(map[reflect.Type]*reflectInfo, len(reflectMap)+1)
+	for k, v := range reflectMap {
+		m[k] = v
 	}
+	m[t] = rm
+	reflectMap = m
+	return rm
+}
 
-	return field, nil
+var reflectMap map[reflect.Type]*reflectInfo
+
+type reflectInfo struct {
+	names map[string][]int
 }