@serejja 2015-09-04
Hi @crast, thanks for submitting this PR! Let me know if you need to discuss some further optimizations or need some help with them please. I'll be willing to merge PRs like that. Thanks!